+++ src/App.js
@@ -1,8 +1,8 @@
 import React from 'react';
-import Container from './components/Container';
+import Header from './components/Header';
 
 function App() {
-  return <Container />;
+  return <Header />;
 }
 
 export default App;
