--- src/components/Container.js
@@ -1,38 +1,0 @@
-import React from 'react';
-import Header from './Header';
-import Nav from './Nav';
-import Footer from './Footer'
-
-// function to check value of currentPage (from useState)
-function Container() {
-  const [currentPage, setCurrentPage] = useState('Home');
-
-// return current page
-  const renderPage = () => {
-    if (currentPage === 'Home') {
-      return <Home />;
-    }
-    if (currentPage === 'About') {
-      return <About />;
-    }
-    if (currentPage === 'Blog') {
-      return <Blog />;
-    }
-    return <Contact />;
-  };
-
-
-  const handlePageChange = (page) => setCurrentPage(page);
-
-  return (
-    <div>
-      {/* We are passing the currentPage from state and the function to update it */}
-      <NavTabs currentPage={currentPage} handlePageChange={handlePageChange} />
-      {/* Here we are calling the renderPage method which will return a component  */}
-      {renderPage()}
-    </div>
-  );
-}
-
-
-export default Container;
